fix wranings

+++ ProcessImageApp/ProcessImageApp/Helpers/UIImage+Effects.m
@@ -176,7 +176,7 @@ CGFloat DegreesToRadians(CGFloat degrees) {return degrees * M_PI / 180;};
 {
   const CGFloat EffectColorAlpha = 0.6;
   UIColor *effectColor = tintColor;
-  int componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
+  long componentCount = CGColorGetNumberOfComponents(tintColor.CGColor);
   if (componentCount == 2) {
     CGFloat b;
     if ([tintColor getWhite:&b alpha:NULL]) {
@@ -238,7 +238,7 @@ CGFloat DegreesToRadians(CGFloat degrees) {return degrees * M_PI / 180;};
     if (hasBlur) {
  
       CGFloat inputRadius = blurRadius * [[UIScreen mainScreen] scale];
-      NSUInteger radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
+      int radius = floor(inputRadius * 3. * sqrt(2 * M_PI) / 4 + 0.5);
       if (radius % 2 != 1) {
         radius += 1; // force radius to be odd so that the three box-blur methodology works.
       }
